scaffold @hive/clickhouse and relocate the sql builder
