+++ packages/services/api/package.json
@@ -27,6 +27,7 @@
     "@graphql-inspector/core": "7.1.3",
     "@graphql-tools/merge": "9.1.1",
     "@hive/cdn-script": "workspace:*",
+    "@hive/clickhouse": "workspace:*",
     "@hive/postgres": "workspace:*",
     "@hive/pubsub": "workspace:*",
     "@hive/schema": "workspace:*",
