+++ tsconfig.json
@@ -25,6 +25,7 @@
     "moduleResolution": "node",
     "strict": true,
     "paths": {
+      "@hive/clickhouse": ["./packages/internal/clickhouse/src/index.ts"],
       "@hive/postgres": ["./packages/internal/postgres/src/index.ts"],
       "@hive/api": ["./packages/services/api/src/index.ts"],
       "@hive/api/*": ["./packages/services/api/src/*"],
